split login and registration pages
