+++ client/.stylelintcache;
@@ -1,1 +1,1 @@
-[{"C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\index.scss":"1","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\App.scss":"2","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\BasketList\\BasketList.module.scss":"3","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Button\\ButtonComponent.module.scss":"4","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\CardList\\CardList.module.scss":"5","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\LoginIcon\\LoginIcon.module.scss":"6","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Modal\\Modal.module.scss":"7","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\ProductsCounter\\ProductsCounter.module.scss":"8","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\BasketListItem\\BasketListItem.module.scss":"9","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\OrderForm\\OrderForm.module.scss":"10","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Pages\\Catalog\\Catalog.module.scss":"11","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Pages\\CurrentProduct\\CurrentProduct.module.scss":"12","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Pages\\SerachList\\SearchList.module.scss":"13"},{"size":273,"mtime":1667905067517,"hashOfConfig":"14"},{"size":650,"mtime":1666368701242,"hashOfConfig":"14"},{"size":47,"mtime":1667921765908,"hashOfConfig":"14"},{"size":50,"mtime":1667905084538,"hashOfConfig":"14"},{"size":103,"mtime":1669464642717,"hashOfConfig":"14"},{"size":101,"mtime":1670505345491,"hashOfConfig":"14"},{"size":0,"mtime":1666368701263,"hashOfConfig":"14"},{"size":189,"mtime":1667921915006,"hashOfConfig":"14"},{"size":0,"mtime":1666368701248,"hashOfConfig":"14"},{"size":357,"mtime":1667921765924,"hashOfConfig":"14"},{"size":88,"mtime":1670495313887,"hashOfConfig":"14"},{"size":916,"mtime":1669476317676,"hashOfConfig":"14"},{"size":127,"mtime":1670857014218,"hashOfConfig":"14"},"1rd9jx2"]
+[{"C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\index.scss":"1","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\App.scss":"2","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\BasketList\\BasketList.module.scss":"3","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Button\\ButtonComponent.module.scss":"4","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\CardList\\CardList.module.scss":"5","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\LoginIcon\\LoginIcon.module.scss":"6","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Modal\\Modal.module.scss":"7","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\ProductsCounter\\ProductsCounter.module.scss":"8","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\BasketListItem\\BasketListItem.module.scss":"9","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\OrderForm\\OrderForm.module.scss":"10","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Pages\\Catalog\\Catalog.module.scss":"11","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Pages\\CurrentProduct\\CurrentProduct.module.scss":"12","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Pages\\SerachList\\SearchList.module.scss":"13","C:\\Users\\AnnaBook\\Desktop\\Bakery-Project\\client\\src\\components\\Pages\\Registration\\Registration.module.scss":"14"},{"size":273,"mtime":1667905067517,"hashOfConfig":"15"},{"size":650,"mtime":1666368701242,"hashOfConfig":"15"},{"size":47,"mtime":1667921765908,"hashOfConfig":"15"},{"size":50,"mtime":1667905084538,"hashOfConfig":"15"},{"size":103,"mtime":1669464642717,"hashOfConfig":"15"},{"size":101,"mtime":1670505345491,"hashOfConfig":"15"},{"size":0,"mtime":1666368701263,"hashOfConfig":"15"},{"size":189,"mtime":1667921915006,"hashOfConfig":"15"},{"size":0,"mtime":1666368701248,"hashOfConfig":"15"},{"size":357,"mtime":1667921765924,"hashOfConfig":"15"},{"size":88,"mtime":1670495313887,"hashOfConfig":"15"},{"size":916,"mtime":1669476317676,"hashOfConfig":"15"},{"size":127,"mtime":1670857014218,"hashOfConfig":"15"},{"size":0,"mtime":1671443700434,"hashOfConfig":"15"},"1rd9jx2"]
