+++ client/src/components/Navigation/NavigationIconMenu.js
@@ -2,7 +2,7 @@ import { NavLink } from 'react-router-dom';
 
 function NavigationIconMenu({
 
-  basket, searchList, main, style, account,
+  basket, searchList, main, style, login,
 
 }) {
   return (
@@ -19,7 +19,7 @@ function NavigationIconMenu({
           <NavLink to="/basket">{basket}</NavLink>
         </li>
         <li className={`${style}`}>
-          <NavLink to="/account">{account}</NavLink>
+          <NavLink to="/login">{login}</NavLink>
         </li>
       </ul>
     </nav>
